webkit/data/layout_tests/chrome/fast/dom/typecheck.html

<html>
  <head></head>
  <body>

<p>
This test checks the behavior of dom operations when called on non-dom
or incompatible receivers with non-dom or incompatible arguments.
</p>

    <script>
if (window.layoutTestController) layoutTestController.dumpAsText();

function addResult(str) {
  var option = document.createElement("option");
  option.innerHTML = str;
  document.getElementById("results").appendChild(option);
}

function trySuspect(name, fun) {
  try {
    result = fun();
    document.write(name + " = " + result + "<br />");
  } catch (e) {
    document.write(name + ": " + e + "<br />");
  }
}

var aDOMImplementation = document.implementation;
var aSelect = document.createElement("select");
var anOption = document.createElement("option");

document.write("DOMImplementation: " + aDOMImplementation + "<br />");
var aNode = document.createElement("div");
var aSecondNode = document.createElement("div");
aNode.appendChild(aSecondNode);

trySuspect("Node::appendChild(DOMImplementation)", function () {
  return aNode.appendChild(aDOMImplementation);
});

trySuspect("Node::appendChild(String)", function () {
  return aNode.appendChild("knort");
});

trySuspect("Node::appendChild(undefined)", function () {
  return aNode.appendChild(void 0);
});

trySuspect("Node::isSameNode(DOMImplementation)", function () {
  return aNode.isSameNode(aDOMImplementation);
});

trySuspect("Node::isSameNode(String)", function () {
  return aNode.isSameNode("foo");
});

trySuspect("Node::isSameNode(undefined)", function () {
  return aNode.isSameNode(void 0);
});

trySuspect("Node::lookupPrefix(DOMImplementation)", function () {
  return aNode.lookupPrefix(aDOMImplementation);
});

trySuspect("Node::lookupPrefix(undefined)", function () {
  return aNode.lookupPrefix(void 0);
});

trySuspect("Node::cloneNode(DOMImplementation)", function () {
  return aNode.cloneNode(aDOMImplementation);
});

trySuspect("Select::add(DOMImplementation, DOMImplementation)", function () {
  return aSelect.add(aDOMImplementation, aDOMImplementation);
});

trySuspect("Select::add(DOMImplementation, Option)", function () {
  return aSelect.add(aDOMImplementation, anOption);
});

trySuspect("Select::add(Option, DOMImplementation)", function () {
  return aSelect.add(anOption, aDOMImplementation);
});

trySuspect("Select::add(undefined, undefined)", function () {
  return aSelect.add(void 0, void 0);
});

trySuspect("Select::add(undefined, Option)", function () {
  return aSelect.add(void 0, anOption);
});

trySuspect("Select::add(Option, undefined)", function () {
  return aSelect.add(anOption, void 0);
});

    </script>
  </body>
</html>
