webkit/data/layout_tests/chrome/fast/dom/typecheck-expected.txt

This test checks the behavior of dom operations when called on non-dom or incompatible receivers with non-dom or incompatible arguments.

DOMImplementation: [object DOMImplementation]
Node::appendChild(DOMImplementation): Error: NOT_FOUND_ERR: DOM Exception 8
Node::appendChild(String): Error: NOT_FOUND_ERR: DOM Exception 8
Node::appendChild(undefined): Error: NOT_FOUND_ERR: DOM Exception 8
Node::isSameNode(DOMImplementation) = false
Node::isSameNode(String) = false
Node::isSameNode(undefined) = false
Node::lookupPrefix(DOMImplementation) = null
Node::lookupPrefix(undefined) = null
Node::cloneNode(DOMImplementation) = [object HTMLDivElement]
Select::add(DOMImplementation, DOMImplementation) = undefined
Select::add(DOMImplementation, Option) = undefined
Select::add(Option, DOMImplementation) = undefined
Select::add(undefined, undefined) = undefined
Select::add(undefined, Option) = undefined
Select::add(Option, undefined) = undefined
